projects/age-detector/requirements.txt
deepface>=0.0.99
gradio>=4.0.0
opencv-python>=4.5.0
pillow>=5.0.0
numpy>=1.14.0
